UI almost ready + :octocat: Added .gitignore
The UI is waiting that the decoder and encoder are ready

OK button not here
The decoder/encoder makes me rage
How to put the file you selected with browse directly after the button, and if not opened, write "Choose a file"?
How to put a watermark on a textbox?